Uploading Side Board
Uploading Side Board. The side board is a deck where the player can click a button to be shown 3 cards from the deck. Then they can select a card to add to the hand

Uploading SideboardDeckManager that takes the top3 cards from the deck and displays them. also handles card being selected
Uploading ClickAddToHand that handles the shown card being selected (clicked)

Updating DeckManager to allow cards to be added to the deck and removed from the deck
Updating BaseSpell to not setup effects if the player manager is null
Updating PlayerManager prefab with sideboard deck
Updating MLM_Toybox to work with sideboard
Updating Test Game Canvas prefab with side board button